Conflict

+++ joern-cli/frontends/ghidra2cpg/src/main/scala/io/joern/ghidra2cpg/Ghidra2Cpg.scala
@@ -16,55 +16,55 @@ import io.joern.ghidra2cpg.passes._
 import io.joern.ghidra2cpg.passes.arm.ArmFunctionPass
 import io.joern.ghidra2cpg.passes.mips.{LoHiPass, MipsFunctionPass}
 import io.joern.ghidra2cpg.passes.x86.{ReturnEdgesPass, X86FunctionPass}
-import io.joern.x2cpg.X2Cpg
-import io.joern.x2cpg.passes.frontend.{MetaDataPass, TypeNodePass}
+import io.joern.x2cpg.X2Cpg.withNewEmptyCpg
+import io.joern.x2cpg.{X2Cpg, X2CpgFrontend}
 import io.shiftleft.codepropertygraph.Cpg
-import io.shiftleft.codepropertygraph.generated.Languages
 import io.shiftleft.passes.KeyPoolCreator
 import utilities.util.FileUtilities
 
 import java.io.File
 import scala.collection.mutable
 import scala.jdk.CollectionConverters._
+import scala.util.Try
 
-class Ghidra2Cpg() {
+class Ghidra2Cpg extends X2CpgFrontend[Config] {
 
-  /** Create a CPG representing the given input file. The CPG is stored at the given output file. The caller must close
-    * the CPG.
-    */
-  def createCpg(inputFile: File, outputFile: Option[String]): Cpg = {
+  override def createCpg(config: Config): Try[Cpg] = {
+    if (config.inputPaths.size != 1) {
+      throw new RuntimeException("This frontend requires exactly one input path")
+    }
 
+    val inputFile = new File(config.inputPaths.head)
     if (!inputFile.isDirectory && !inputFile.isFile) {
       throw new InvalidInputException(s"$inputFile is not a valid directory or file.")
     }
 
-    val cpg = X2Cpg.newEmptyCpg(outputFile)
+    withNewEmptyCpg(config.outputPath, config) { (cpg, _) =>
+      better.files.File.usingTemporaryDirectory("ghidra2cpg_tmp") { tempWorkingDir =>
+        initGhidra()
+        val locator = new ProjectLocator(tempWorkingDir.path.toAbsolutePath.toString, CommandLineConfig.projectName)
+        var program: Program = null
+        var project: Project = null
 
-    better.files.File.usingTemporaryDirectory("ghidra2cpg_tmp") { tempWorkingDir =>
-      initGhidra()
-      val locator = new ProjectLocator(tempWorkingDir.path.toAbsolutePath.toString, CommandLineConfig.projectName)
-      var program: Program = null
-      var project: Project = null
-
-      try {
-        val projectManager = new HeadlessGhidraProjectManager
-        project = projectManager.createProject(locator, null, false)
-        program = AutoImporter.importByUsingBestGuess(inputFile, null, this, new MessageLog, TaskMonitor.DUMMY)
-        addProgramToCpg(program, inputFile.getCanonicalPath, cpg)
-      } catch {
-        case e: Exception =>
-          e.printStackTrace()
-      } finally {
-        if (program != null) {
-          AutoAnalysisManager.getAnalysisManager(program).dispose()
-          program.release(this)
+        try {
+          val projectManager = new HeadlessGhidraProjectManager
+          project = projectManager.createProject(locator, null, false)
+          program = AutoImporter.importByUsingBestGuess(inputFile, null, this, new MessageLog, TaskMonitor.DUMMY)
+          addProgramToCpg(program, inputFile.getAbsolutePath, cpg)
+        } catch {
+          case e: Exception =>
+            e.printStackTrace()
+        } finally {
+          if (program != null) {
+            AutoAnalysisManager.getAnalysisManager(program).dispose()
+            program.release(this)
+          }
+          project.close()
+          FileUtilities.deleteDir(locator.getProjectDir)
+          locator.getMarkerFile.delete
         }
-        project.close()
-        FileUtilities.deleteDir(locator.getProjectDir)
-        locator.getMarkerFile.delete
       }
     }
-    cpg
   }
 
   private def initGhidra(): Unit = {
@@ -115,31 +115,50 @@ class Ghidra2Cpg() {
       .map(x => x.getAddress().getOffset -> x.getValue.toString)
       .toMap
 
-    new MetaDataPass(cpg, Languages.GHIDRA).createAndApply()
-    new NamespacePass(cpg, flatProgramAPI.getProgramFile).createAndApply()
+    // We touch every function twice, regular ASM and PCode
+    // Also we have + 2 for MetaDataPass and NamespacePass
+    val numOfKeypools   = functions.size * 3 + 2
+    val keyPoolIterator = KeyPoolCreator.obtain(numOfKeypools).iterator
+
+    new MetaDataPass(fileAbsolutePath, cpg).createAndApply()
+    new NamespacePass(cpg, fileAbsolutePath, keyPoolIterator.next()).createAndApply()
 
     program.getLanguage.getLanguageDescription.getProcessor.toString match {
       case "MIPS" =>
-        new MipsFunctionPass(program, address2Literals, fileAbsolutePath, functions, cpg, decompiler).createAndApply()
-        new LoHiPass(cpg).createAndApply()
+        functions.foreach { function =>
+          new MipsFunctionPass(
+            program,
+            address2Literals,
+            fileAbsolutePath,
+            function,
+            cpg,
+            keyPoolIterator.next(),
+            decompiler
+          ).createAndApply()
+          new LoHiPass(cpg).createAndApply()
+        }
       case "AARCH64" | "ARM" =>
-        new ArmFunctionPass(program, fileAbsolutePath, functions, cpg, decompiler)
-          .createAndApply()
+        functions.foreach { function =>
+          new ArmFunctionPass(program, fileAbsolutePath, function, cpg, keyPoolIterator.next(), decompiler)
+            .createAndApply()
+        }
       case _ =>
-        new X86FunctionPass(program, fileAbsolutePath, functions, cpg, decompiler)
-          .createAndApply()
-        new ReturnEdgesPass(cpg).createAndApply()
+        functions.foreach { function =>
+          new X86FunctionPass(program, fileAbsolutePath, function, cpg, keyPoolIterator.next(), decompiler)
+            .createAndApply()
+          new ReturnEdgesPass(cpg).createAndApply()
+        }
     }
-
-    new TypeNodePass(Types.types.toList, cpg)
-    new JumpPass(cpg).createAndApply()
-    new LiteralPass(cpg, flatProgramAPI).createAndApply()
+    new TypesPass(cpg).createAndApply()
+    new JumpPass(cpg, keyPoolIterator.next()).createAndApply()
+    new LiteralPass(cpg, address2Literals, program, flatProgramAPI, keyPoolIterator.next()).createAndApply()
   }
 
   private class HeadlessProjectConnection(projectManager: HeadlessGhidraProjectManager, connection: GhidraURLConnection)
       extends DefaultProject(projectManager, connection) {}
 
   private class HeadlessGhidraProjectManager extends DefaultProjectManager {}
+
 }
 
 object Types {
